Cartier UAE Chatbot/requirements.txt
google-genai
chromadb
python-dotenv
pytest
colorama
